requirements.txt
asgiref==3.11.1
click==8.3.2
croniter==6.2.2
Django==6.0.4
django-redis==6.0.0
django-rq==4.1.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.3.1
gunicorn==25.3.0
packaging==26.1
Pillow==11.0.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
rq==2.8.0
six==1.17.0
sqlparse==0.5.5
whitenoise==6.12.0